Adding in new 'cgz' scanner.
CGData Ziped files (.cgz) are single zipped files that represent a
collection of CGData files. This make it easier to move collections
of dependent files as a single unit, as well as making the whole
repo 'read-only', to prevent causual user changes to the meta-info
structure.

The compiler will now also scan .cgz files it finds in the specified
directories.