Change to add data unloading during compile.
Added the 'unload' method to base classes. This give the compiler
the option of unloaded information after using it. This is so
when the compiler visits all the different objects in a library
it doesn't end up with a copy of every one of them in memory.
This method is only meant to unload the core data, not the meta-data.
If the path variable is set, the object should be able to reload.